Updated whoogle + jackett + other stuff

+++ radarr/rootfs/etc/cont-init.d/20-folders.sh
@@ -12,15 +12,15 @@ if [ ! -d /share/downloads ]; then
   chown -R abc:abc /share/downloads
 fi
 
-if [ ! -d /config/addons_config/radarr ]; then
-if [ -d /config/radarr ]; then
+if [ -d /config/radarr ] && [ ! -d /config/addons_config/radarr ]; then
   echo "Moving to new location /config/addons_config/radarr"
   mkdir -p /config/addons_config/radarr
   chown -R abc:abc /config/addons_config/radarr
   mv /config/radarr/* /config/addons_config/radarr/
   rm -r /config/radarr
 fi
- 
+
+if [ ! -d /config/addons_config/radarr ]; then
   echo "Creating /config/addons_config/radarr"
   mkdir -p /config/addons_config/radarr
   chown -R abc:abc /config/addons_config/radarr
